==> notebooks/2 - build_model/1 - feature_selection.ipynb <==
import pandas as pd
from tqdm import tqdm
from copy import copy
import statsmodels.api as sm

from sklearn.feature_selection import SelectKBest, f_classif

data = pd.read_parquet('../../data/curated/clean_stats_13-22.parquet')

data.columns[18:28]

data.columns[78:131]

data.columns[176:]

# Feature Selection

potential_features = [
    'is_captain', 'player_position_defender', 'player_position_rover',
    'player_position_key_defender', 'player_position_key_forward',
    'player_position_forward', 'player_position_INT',
    'player_position_ruck', 'player_position_sub', 'player_position_wing',
    'rating_points', 'winning_margin', 'won_match', '30_and_2',
    'high_goal_scorer','top_disposal_getter', 'top_clearance_getter',
    'midfielder_goals',
    'kicks_proportion', 'marks_proportion',
    'handballs_proportion', 'disposals_proportion',
    'effective_disposals_proportion', 'goals_proportion',
    'afl_fantasy_score_proportion', 'SC_proportion', 'behinds_proportion',
    'hitouts_proportion', 'tackles_proportion', 'rebounds_proportion',
    'inside_fifties_proportion', 'clearances_proportion',
    'clangers_proportion', 'free_kicks_for_proportion',
    'free_kicks_against_proportion', 'contested_possessions_proportion',
    'uncontested_possessions_proportion', 'contested_marks_proportion',
    'marks_inside_fifty_proportion', 'one_percenters_proportion',
    'bounces_proportion', 'goal_assists_proportion',
    'centre_clearances_proportion', 'stoppage_clearances_proportion',
    'score_involvements_proportion', 'metres_gained_proportion',
    'turnovers_proportion', 'intercepts_proportion',
    'tackles_inside_fifty_proportion', 'contest_def_losses_proportion',
    'contest_def_one_on_ones_proportion',
    'contest_off_one_on_ones_proportion', 'contest_off_wins_proportion',
    'def_half_pressure_acts_proportion', 'effective_kicks_proportion',
    'f50_ground_ball_gets_proportion', 'ground_ball_gets_proportion',
    'hitouts_to_advantage_proportion', 'intercept_marks_proportion',
    'marks_on_lead_proportion', 'score_launches_proportion',
    'shots_at_goal_proportion', 'spoils_proportion',
    'coaches_votes', 'average_votes_prev'
    ]
target = 'brownlow_votes'

# Stepwise Selection

Define functions forwards and backward that conduct forward selection and backward selection respectively.

Then iterate through a stepwise selection process until the model AIC no longer decreases.

def forward(current_features, X, y_train, min_aic=10_000_000):
    
    next_feature = None
        
    for f in tqdm(potential_features, desc="Forward Selection"):

        # do not want to add same feature multiple times
        if f not in current_features:
            
            curr_features = list(current_features) + [f]


            # train model with new feature added
            X_train = X[curr_features]
            
            model = sm.MNLogit(y_train, sm.add_constant(X_train)).fit(disp=0)

            # calc aic
            local_aic = model.aic


            if local_aic < min_aic:

                next_feature = f

                min_aic = local_aic  
    
    return (min_aic, next_feature)

def backward(current_features, X, y_train, min_aic=10_000_000):
    
    next_feature = None
        
    for f in tqdm(potential_features, desc="Backward Selection"):

        # do not want to add same feature multiple times
        if f in current_features:
            
            curr_features = list(current_features)
            curr_features.remove(f)


            # train model with new feature added
            X_train = X[curr_features]
            
            model = sm.MNLogit(y_train, sm.add_constant(X_train)).fit(disp=0)

            # calc aic
            local_aic = model.aic


            if local_aic < min_aic:

                next_feature = f

                min_aic = local_aic  
    
    return (min_aic, next_feature)

# # since there are many features, this takes a long time to run.
# # features are saved to features.txt for loading 

# current_features = []
# previous_aic = float('inf')  # Start with a large AIC
# X = data.query('season < 2022')[potential_features]
# y_train = data.query('season < 2022')[target]

# while True:
#     # Forward selection
#     forward_aic, forward_feature = forward(current_features, X, y_train, previous_aic)
    
#     # Backward elimination
#     backward_aic, backward_feature = backward(current_features, X, y_train, previous_aic)
    
#     # Check which method (forward or backward) improves the model more
#     if forward_aic < backward_aic and forward_aic < previous_aic:
#         current_features.append(forward_feature)
#         print(f'added: {forward_feature}\n')
#         previous_aic = forward_aic
#     elif backward_aic < previous_aic:
#         current_features.remove(backward_feature)
#         print(f'removed: {backward_feature}\n')
#         previous_aic = backward_aic
#     else:
#         # If no improvement is made, break out of the loop
#         print('stepwise selection finished')
#         break

# # write features to text file to avoid stepwise selection when making changes
# with open('stepwise_selection_features.txt', 'w') as f:
#     for feature in current_features:
#         f.write(f'{feature}\n')

# SelectKBest feature selection

X_train, y_train = data.query('season < 2022')[potential_features], data.query('season < 2022')[target]

k_best = SelectKBest(score_func=f_classif, k=30)  # Select 10 best features

k_best.fit_transform(X_train, y_train)
current_features = list(k_best.get_feature_names_out())

current_features

# features that have been dropped during stepwise selection
dropped_features = set(potential_features) - set(current_features)
dropped_features

# add and remove features based on domain knowledge statistical significance
# explore whether features improve model performance

feat = list(set(current_features) - set([
   'effective_kicks_proportion',
   'shots_at_goal_proportion',
   'clearances_proportion',
   'effective_disposals_proportion',
   'player_position_rover',
   'rating_points',
   'disposals_proportion',
   'marks_proportion',
   'score_launches_proportion',
   'turnovers_proportion',
   'stoppage_clearances_proportion',
   'inside_fifties_proportion',
   'score_involvements_proportion',
   'ground_ball_gets_proportion',
   'top_clearance_getter',
   'uncontested_possessions_proportion',
   '30_and_2'
   
])) 

feat += [
    'winning_margin',
    'is_captain',
    'hitouts_to_advantage_proportion',
    'intercept_marks_proportion',
    'spoils_proportion',
    'contested_marks_proportion',

    # 'bounces_proportion',
]

len(feat)

X_train, y_train = data.query('season < 2022')[feat], data.query('season < 2022')[target]

model2 = sm.MNLogit(y_train, sm.add_constant(X_train)).fit(disp=0, maxiter=10_000)
model2.aic

model2.summary()

# final features are:
feat

# write features to text file so they can be opened in next notebook
with open('final_features.txt', 'w') as f:
    for feature in feat:
        f.write(f'{feature}\n')
