==> notebooks/2 - build_model/2 - build_model.ipynb <==
import pandas as pd
import statsmodels.api as sm

data = pd.read_parquet('../../data/curated/clean_stats_13-22.parquet')
data.set_index('player_id', inplace=True)

features = []
with open('final_features.txt', 'r') as f:
    for line in f.readlines():
        features.append(line.strip())

target = 'brownlow_votes'

X_train, y_train = data.query('season < 2022')[features], data.query('season < 2022')[target]
X_test, y_test = data.query('season == 2022')[features], data.query('season == 2022')[target]

model = sm.MNLogit(y_train, sm.add_constant(X_train)).fit(disp=0, maxiter=10_000)

predictions = model.predict(sm.add_constant(X_test))

df_2022 = data.query('season == 2022')[['match_id', 'match_round', 'player_first_name', 'player_last_name', 'player_team']]

out_df = pd.concat([df_2022, predictions], axis=1)

out_df

out_df.to_parquet('../../data/curated/2022_vote_probs.parquet')
