==> notebooks/2 - build_model/3 - normalise_probabilities.ipynb <==
import pandas as pd

df = pd.read_parquet('../../data/curated/2022_vote_probs.parquet')

df

match_ids = df.match_id.unique()

def standardise_votes(df):
    
    ini_sum_1 = sum(df['1'])
    ini_sum_2 = sum(df['2'])
    ini_sum_3 = sum(df['3'])
    
    df['new_3'] = df['3'] / ini_sum_3
    df['new_2'] = df['2'] / ini_sum_2
    df['new_1'] = df['1'] / ini_sum_1

    df['new_0'] = 1 - (df['new_1'] + df['new_2'] + df['new_3'])
    
    df['exp_votes'] = 1*df['new_1'] + 2*df['new_2'] + 3*df['new_3']
    
    
    return df[['player_id', 'match_id', 'player_first_name', 'player_last_name', 'player_team', 'new_0', 'new_1', 'new_2', 'new_3', 'exp_votes']]

final_df = pd.DataFrame(columns=['player_id', 'match_id', 'player_first_name', 'player_last_name', 'player_team', 'new_0', 'new_1', 'new_2', 'new_3', 'exp_votes'])


for idx in match_ids:
    
    df_match = df[df['match_id'] == idx].reset_index() 
    
    final_df = pd.concat([final_df, standardise_votes(df_match)])

final_df

final_df.to_parquet('../../data/curated/adjusted_probabilities_2022.parquet')
